+++ src/test/encrypt/encrypt-scheme_CKKS-poly_4096-sec_256-type_float64.test.js
@@ -28,22 +28,22 @@ describe.skip('encrypt on CKKS', () => {
       expect(context.parametersSet).toBe(true)
 
       const encoder = Morfix.CKKSEncoder({
-        context: context
+        context
       })
 
       const keyGenerator = Morfix.KeyGenerator({
-        context: context
+        context
       })
 
       const publicKey = keyGenerator.getPublicKey()
       const secretKey = keyGenerator.getSecretKey()
       const encryptor = Morfix.Encryptor({
-        context: context,
-        publicKey: publicKey
+        context,
+        publicKey
       })
       const decryptor = Morfix.Decryptor({
-        context: context,
-        secretKey: secretKey
+        context,
+        secretKey
       })
 
       // Create data to be encrypted
@@ -51,27 +51,20 @@ describe.skip('encrypt on CKKS', () => {
         length: 2048
       }).map((x, i) =>  i)
 
-      // Create a plainText variable and encode the vector to it
-      const plainText = Morfix.PlainText()
-
-      encoder.encode({
+      // Encode the Array
+      const plainText = encoder.encode({
         array,
-        scale: Math.pow(2, 58),
+        scale: Math.pow(2, 58)
+      })
+
+      // Encrypt the PlainText
+      const cipherText = encryptor.encrypt({
         plainText
       })
 
-      // Create a cipherText variable and encrypt the plainText to it
-      const cipherText = Morfix.CipherText()
-      encryptor.encrypt({
-        plainText,
+      // Decrypt the CipherText
+      const decryptedPlainText = decryptor.decrypt({
         cipherText
-      })
-
-      // Create a new plainText variable to store the decrypted cipherText
-      const decryptedPlainText = Morfix.PlainText()
-      decryptor.decrypt({
-        cipherText,
-        plainText: decryptedPlainText
       })
 
       // Decode the PlainText