+++ src/test/galoiskey/galoiskey-scheme_CKKS-poly_16384-sec_128.test.js
@@ -28,7 +28,7 @@ describe.skip('galoiskey on CKKS', () => {
       expect(context.parametersSet).toBe(true)
 
       const keyGenerator = Morfix.KeyGenerator({
-        context: context
+        context
       })
 
       const spyGenGaloisKeys = jest.spyOn(keyGenerator, 'genGaloisKeys')