+++ src/test/keypair/keypair-scheme_CKKS-poly_4096-sec_256.test.js
@@ -28,7 +28,7 @@ describe.skip('keypair on CKKS', () => {
       expect(context.parametersSet).toBe(true)
 
       const keyGenerator = Morfix.KeyGenerator({
-        context: context
+        context
       })
 
       const spyGetSecretKey = jest.spyOn(keyGenerator, 'getSecretKey')
@@ -43,7 +43,7 @@ describe.skip('keypair on CKKS', () => {
       const spySaveSecretKey = jest.spyOn(secretKey, 'save')
       const secretKeyBase64 = secretKey.save()
       expect(spySaveSecretKey).toHaveBeenCalled()
-      
+
       const spySavePublicKey = jest.spyOn(publicKey, 'save')
       const publicKeyBase64 = publicKey.save()
       expect(spySavePublicKey).toHaveBeenCalled()